fix: complete int-to-decimal migration for rating system
Change all rating-related types from int to decimal (0.0-10.0 scale):
- ActivityFeedDTOs.cs: RatingScore changed to decimal?
- ActivityFeedService.cs: Parse ratings as decimal from JSON metadata
- ActivityFeed.razor: Update GetRatingColorHex to use decimal comparisons
- RatingsController.cs: UpsertRatingRequest Score changed to decimal
- FoodItem.cs: Ratings dictionary changed to decimal?
- FoodService.cs/IFoodService.cs: UpdateRatingAsync accepts decimal
- FoodDetails.razor, AddFood.razor, CookBook.razor: Fixed slider bindings
- FriendRatings.razor: Format score display with decimal format
- UserRecipeService.cs: Use decimal literals in Math.Clamp
- AddMemberToCollectionDialog.razor: Add missing ApiClient using

+++ StorhaugenEats.API/Controllers/RatingsController.cs
@@ -167,4 +167,4 @@ public class RatingsController : ControllerBase
 
 }
 
-public record UpsertRatingRequest(Guid GlobalRecipeId, int Score, string? Comment);
+public record UpsertRatingRequest(Guid GlobalRecipeId, decimal Score, string? Comment);
